Adding sample repository for firestore DB storage
