+++ app/src/main/java/net/aiscope/gdd_app/model/SampleModels.kt
@@ -1,7 +1,5 @@
 package net.aiscope.gdd_app.model
 
-//import com.google.firebase.Timestamp
-//import com.google.firebase.firestore.FirebaseFirestore
 import net.aiscope.gdd_app.extensions.replaceElementAt
 import java.io.File
 import java.util.Calendar
@@ -33,16 +31,6 @@ data class Sample(
     fun nextImageName(): String = "${id}_image_${captures.completedCaptureCount()}"
 
     fun nextMaskName(): String = "${id}_mask_${captures.completedCaptureCount()}"
-
-    /*
-    fun buildSampleCollection() = SampleCollection(
-        createdOn = Timestamp(createdOn.time),
-        uploadedBy = FirebaseFirestore.getInstance().
-        collection("microscopist").document(microscopist),
-        location = id,
-        numberOfImages = captures.completedCaptureCount()
-    )
-     */
 }
 
 enum class SampleStatus(val id: Short) {
