--- app/src/main/java/net/aiscope/gdd_app/network/FirestoreUtil.kt
@@ -1,17 +1,0 @@
-package net.aiscope.gdd_app.network
-//import com.google.firebase.firestore.FirebaseFirestore
-
-class FirestoreUtil {
-    /*
-    object FirestoreUtil {
-        private var FIRESTORE: FirebaseFirestore? = null
-        val firestore: FirebaseFirestore?
-            get() {
-                if (FIRESTORE == null) {
-                    FIRESTORE = FirebaseFirestore.getInstance()
-                }
-                return FIRESTORE
-            }
-    }
-     */
-}
