+++ app/src/main/java/net/aiscope/gdd_app/network/RemoteStorage.kt
@@ -6,7 +6,6 @@ import androidx.work.NetworkType
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.workDataOf
-//import com.google.firebase.firestore.FirebaseFirestore
 import net.aiscope.gdd_app.model.Sample
 
 interface RemoteStorage {
@@ -27,14 +26,4 @@ interface RemoteStorage {
 
         WorkManager.getInstance(context).enqueue(uploadWorkRequest)
     }
-
-    /*
-    fun storeSampleCollection(sample: Sample) {
-        val sampleCollection = sample.buildSampleCollection()
-        FirebaseFirestore.getInstance()
-            .collection("samples")
-            .document(sampleCollection.location)
-            .set(sampleCollection)
-    }
-     */
 }
